+++ linux-client/linux-client.vcxproj
@@ -35,9 +35,9 @@
     </ProjectConfiguration>
   </ItemGroup>
   <PropertyGroup Label="Globals">
-    <ProjectGuid>{5081285b-6c03-4995-bb19-f22cecfb615b}</ProjectGuid>
+    <ProjectGuid>{e03a0032-ce64-4f3b-8275-12353fe0e3cb}</ProjectGuid>
     <Keyword>Linux</Keyword>
-    <RootNamespace>linux_server</RootNamespace>
+    <RootNamespace>linux_client</RootNamespace>
     <MinimumVisualStudioVersion>15.0</MinimumVisualStudioVersion>
     <ApplicationType>Linux</ApplicationType>
     <ApplicationTypeRevision>1.0</ApplicationTypeRevision>
@@ -59,7 +59,6 @@
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'" Label="Configuration">
     <UseDebugLibraries>true</UseDebugLibraries>
-    <UseOfStl>libstdc++_static</UseOfStl>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'" Label="Configuration">
     <UseDebugLibraries>false</UseDebugLibraries>
@@ -72,27 +71,26 @@
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.props" />
   <ImportGroup Label="ExtensionSettings" />
-  <ImportGroup Label="Shared" />
+  <ImportGroup Label="Shared">
+    <Import Project="..\share\share.vcxitems" Label="Shared" />
+  </ImportGroup>
   <ImportGroup Label="PropertySheets" />
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <EnableASAN>true</EnableASAN>
   </PropertyGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\lib\share_lib_server.cpp" />
-    <ClCompile Include="..\server\server.cpp" />
+    <ClCompile Include="..\client\client.cpp" />
+    <ClCompile Include="..\client\share_lib_client.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\lib\share_lib_server.h" />
+    <ClInclude Include="..\client\share_lib_client.h" />
   </ItemGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <ClCompile>
       <CppLanguageStandard>c++17</CppLanguageStandard>
       <AdditionalOptions>-D _DEBUG %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
-    <Link>
-      <LibraryDependencies>pthread;%(LibraryDependencies)</LibraryDependencies>
-    </Link>
   </ItemDefinitionGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets" />
