+++ server/share_lib_server.h
@@ -16,6 +16,7 @@
 #include <ctime>
 #include <filesystem>
 #include <stdarg.h>
+#include "wrapper.h"
 
 #ifdef _WIN32
 #define _WINSOCKAPI_
@@ -163,15 +164,4 @@ int  Dequeue(SOCKET* socket_descriptor, Queue** front, Queue** rear);
 
 bool _endingSequence(char* buffer, unsigned long long size);
 
-////////////////////////////////////////////////////////////////////////////////////
 
-// calloc Wrapper
-void* Calloc(size_t nmemb, size_t size);
-// fprintf Wrapper
-void ShowErr(const char* str);
-// free Wrapper
-void Free(void * arg, size_t size=0);
-// strcpy Wrapper
-void Strcpy(char* dest, size_t size, const char* src);
-// asprintf Wrapper
-int Asprintf(char*& buffer, const char* Format, ...);
