--- windows/server/test.txt
@@ -1,7 +1,0 @@
-zzz
-sss
-qqq
-aaa
-BBB
-ddd
-AAA
